Refactor Win32 stuff
